work on ERC20ForSPL
