+++ ERC20ForSPL/contracts/ERC20ForSPL.sol
@@ -30,7 +30,7 @@ contract ERC20ForSPL is OwnableUpgradeable, UUPSUpgradeable {
         _disableInitializers();
     }
 
-    function initializeParent(bytes32 _tokenMint) public onlyInitializing {
+    function initializeParent(bytes32 _tokenMint) public initializer {
         __Ownable_init(msg.sender);
         if (!SPL_TOKEN.getMint(_tokenMint).isInitialized) revert InvalidTokenMint();
         if (!METAPLEX.isInitialized(_tokenMint)) revert MissingMetaplex();
@@ -189,43 +189,5 @@ contract ERC20ForSPL is OwnableUpgradeable, UUPSUpgradeable {
     function getAccountDelegateData(address who) public view returns(bytes32, uint64) {
         ISPLToken.Account memory account = SPL_TOKEN.getAccount(solanaAccount(who));
         return (account.delegate, account.delegated_amount);
-    }
-}
-
-
-contract ERC20ForSPLMintable is ERC20ForSPL {
-    function initialize(
-        string memory _name,
-        string memory _symbol,
-        uint8 _decimals
-    ) public initializer {       
-        ERC20ForSPL.initializeParent(_initialize(_name, _symbol, _decimals));
-    }
-
-    function findMintAccount() public pure returns (bytes32) {
-        return SPL_TOKEN.findAccount(bytes32(0));
-    }
-
-    function mint(address to, uint256 amount) public onlyOwner {
-        if (to == address(0)) revert EmptyToAddress();
-        if (totalSupply() + amount > type(uint64).max) revert AmountExceedsUint64();
-
-        bytes32 toSolana = solanaAccount(to);
-        if (SPL_TOKEN.isSystemAccount(toSolana)) {
-            SPL_TOKEN.initializeAccount(_salt(to), tokenMint);
-        }
-
-        SPL_TOKEN.mintTo(tokenMint, toSolana, uint64(amount));
-        emit Transfer(address(0), to, amount);
-    }
-
-    function _initialize(
-        string memory _name,
-        string memory _symbol,
-        uint8 _decimals
-    ) private returns (bytes32) {
-        bytes32 mintAddress = SPL_TOKEN.initializeMint(bytes32(0), _decimals);
-        METAPLEX.createMetadata(mintAddress, _name, _symbol, "");
-        return mintAddress;
     }
 }
