+++ ERC20ForSPL/contracts/test/ERC20ForSPLMintableV2.sol
@@ -1,12 +1,12 @@
 // SPDX-License-Identifier: MIT
 pragma solidity 0.8.21;
 
-import './ERC20ForSPL.sol';
+import '../ERC20ForSPLMintable.sol';
 
 // !!!!!!!!!!!!!!!!!
 // !!! This contract shouldn't be used in production, it's made only to validate the upgradibility on Neon network !!!
 // !!!!!!!!!!!!!!!!!
-contract ERC20ForSPLV2 is ERC20ForSPL {
+contract ERC20ForSPLMintableV2 is ERC20ForSPL {
     function getDummyData() public view returns(uint256) {
         return 12345;
     }
