+++ ERC20ForSPL/hardhat.config.js
@@ -38,15 +38,7 @@ module.exports = {
             gas: "auto",
             gasPrice: "auto",
             isFork: true
-        },
-        /* hardhat: {
-            forking: {
-                live: false,
-                saveDeployments: false,
-                accounts: [process.env.PRIVATE_KEY_OWNER, process.env.USER1_KEY, process.env.USER2_KEY, process.env.USER3_KEY],
-                url: "https://devnet.neonevm.org"
-            }
-        } */
+        }
     },
     mocha: {
         timeout: 180000
